Connecting to Spotify and create a mechanism to switch accounts - To be reviewed

+++ database/migrations/2020_04_24_211238_users_accounts.php
@@ -16,13 +16,13 @@ class UsersAccounts extends Migration
         Schema::create('users_accounts', function (Blueprint $table) {
             $table->string('pseudo', 100);
 
-            $table->integer('deezer_id')->unsigned()->nullable();
+            $table->bigInteger('deezer_id')->unsigned()->nullable();
             $table->foreign('deezer_id')->references('id')->on('users');
 
-            $table->integer('spotify_id')->unsigned()->nullable();
+            $table->bigInteger('spotify_id')->unsigned()->nullable();
             $table->foreign('spotify_id')->references('id')->on('users');
 
-            $table->integer('discogs_id')->unsigned()->nullable();
+            $table->bigInteger('discogs_id')->unsigned()->nullable();
             $table->foreign('discogs_id')->references('id')->on('users');
         });
